# Non-contiguous tensor iteration optimization