@mratsim 2024-07-17
> Also, are 0-rank tensors supported? And can they be iterated upon? Right now I added an assert to make sure that the rank is >0.

iirc I tried to make them work and got involved into nasty compiler issues, but that was 7+ years ago. It's fine to assume unsupported for now.