@mratsim 2024-07-17
> By the way, the same optimization can be applied to dualStridedIteration and tripleStridedIteration in the case when the tensor shapes can be broadcast to the same shape. This should bring similar speed-ups to operations like apply2_inline when non-contiguous tensors are involved.

Note that I have a refactoring to allow iteration on a variadic number of tensors here: https://github.com/mratsim/Arraymancer/blob/v0.7.32/src/arraymancer/laser/strided_iteration/foreach_common.nim#L101-L119

Instead of doing the dual/triple it would be more future forward to modify those and then replace the old iterations proc.