@darkestpigeon 2024-07-17
Cool, I'll check out the variadic version. Might take a while, I'm new to nim and I see that the code is macro-heavy.
Also, regarding the iteration. Is it safe to assume that all the tensors we're iterating on have compatible shapes (e.g. can be broadcasted to a common shape)?