@mratsim 2024-07-17
> Is it safe to assume that all the tensors we're iterating on have compatible shapes (e.g. can be broadcasted to a common shape)?

No, we can add a check and fallback to a slow-path.

> Might take a while, I'm new to nim and I see that the code is macro-heavy.

No worries, unfortunately that was iirc the cleanest solution.

The reference code I used while developing generalizing the macros is this one: https://github.com/mratsim/laser/blob/master/benchmarks/loop_iteration/iter05_fusedpertensor.nim#L9-L143