@Vindaar 2024-09-20
This PR caused a small regression related to `reshape` that was not caused, due to the broken CI. I fixed it in #666. The issue was that the new `reshape_with_copy` implementation did not reset the `offset` of the input tensor.